Added initial files for Develop Strategy V2 domain
A new version of the "Develop Strategy" domain has been initiated with the addition of `Run.DevelopStrategy_V2.h` and `Run.DevelopStrategy_V2.cpp` files. These files have been included in the project build process and the project's file structure. The source file contains a `run()` function within the `domain::develop_strategy_V2` namespace, which appears to be the main function for this new domain. The header file declares the same namespace but does not yet declare any functions or classes.